feat: separate models into two files for discussions and messages models
