+++ chats/apps/discussions/models/__init__.py
@@ -1,1 +1,2 @@
-from .models import Discussion, DiscussionMessage, DiscussionUser  # noqa
+from .discussion import Discussion, DiscussionUser  # noqa
+from .discussion_message import DiscussionMessage, DiscussionMessageMedia  # noqa
